% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{list_disciplines}
\alias{list_disciplines}
\title{List all available disciplines}
\usage{
list_disciplines()
}
\value{
A tibble with two columns:
\itemize{
\item name - Character. Discipline name
\item id - Character. Internal API identifier
}
}
\description{
Returns all disciplines (Bilim Dalı values) in the National Thesis Center
database.
Turkish: Bilim Dalı
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
disciplines <- list_disciplines()
head(disciplines)
#> # A tibble: 6 x 2
#>   name                         id
#>   <chr>                        <chr>
#> 1 ACIK DENIZ YAPILARI          1
#> 2 ADLI BILIMLER                2
\dontshow{\}) # examplesIf}
}
\seealso{
Other lookup functions: 
\code{\link{list_divisions}()},
\code{\link{list_institutes}()},
\code{\link{list_subjects}()},
\code{\link{list_universities}()}
}
\concept{lookup functions}
