% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{list_universities}
\alias{list_universities}
\title{List all available universities}
\usage{
list_universities()
}
\value{
A tibble with two columns:
\itemize{
\item name - Character. University name
\item id - Character. Internal API identifier
}
}
\description{
Returns all universities in the National Thesis Center database.
Turkish: Üniversite
}
\examples{
\dontrun{
unis <- list_universities()
}
}
