% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{stats_subjects}
\alias{stats_subjects}
\title{Get thesis statistics by subject}
\usage{
stats_subjects()
}
\value{
A tibble containing thesis counts per subject and type.
}
\description{
Get thesis statistics by subject
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
stats <- stats_subjects()
\dontshow{\}) # examplesIf}
}
\seealso{
Other statistics functions: 
\code{\link{stats_types}()},
\code{\link{stats_universities}()},
\code{\link{stats_years}()}
}
\concept{statistics functions}
