% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{stats_types}
\alias{stats_types}
\title{Get thesis statistics by type}
\usage{
stats_types()
}
\value{
A tibble containing total thesis counts per type.
}
\description{
Get thesis statistics by type
}
\examples{
\dontrun{
stats <- stats_types()
}
}
