% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{stats_years}
\alias{stats_years}
\title{Get thesis statistics by year}
\usage{
stats_years()
}
\value{
A tibble containing thesis counts per year and type.
}
\description{
Get thesis statistics by year
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
stats <- stats_years()
tail(stats)
#> # A tibble: 6 x 4
#>    year yuksek_lisans doktora toplam
#>   <int>         <int>   <int>  <int>
#> 1  2025         22615    4416  27031
\dontshow{\}) # examplesIf}
}
\seealso{
Other statistics functions: 
\code{\link{stats_subjects}()},
\code{\link{stats_types}()},
\code{\link{stats_universities}()}
}
\concept{statistics functions}
