% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tezr-package.R
\docType{package}
\name{tezr-package}
\alias{tezr}
\alias{tezr-package}
\title{tezr: Access the Turkiye's National Thesis Center Database}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access the Turkiye's National Thesis Center (Ulusal Tez Merkezi) database at \url{https://tez.yok.gov.tr}. Search, retrieve, and analyze dissertation metadata from Turkish universities. Supports basic and advanced search, thesis details retrieval, lookup tables, and aggregate statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://eremrah.com/tezr}
  \item \url{https://github.com/emraher/tezr}
  \item Report bugs at \url{https://github.com/emraher/tezr/issues}
}

}
\author{
\strong{Maintainer}: Emrah Er \email{eer@eremrah.com}

}
\keyword{internal}
