% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{list_disciplines}
\alias{list_disciplines}
\title{List all available disciplines}
\usage{
list_disciplines()
}
\value{
A tibble with two columns:
\itemize{
\item name - Character. Discipline name
\item id - Character. Internal API identifier
}
}
\description{
Returns all disciplines (Bilim Dalı values) in the National Thesis Center database.
Turkish: Bilim Dalı
}
\examples{
\dontrun{
disciplines <- list_disciplines()
}
}
