% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{list_divisions}
\alias{list_divisions}
\title{List all available divisions}
\usage{
list_divisions()
}
\value{
A tibble with two columns:
\itemize{
\item name - Character. Division name
\item id - Character. Internal API identifier
}
}
\description{
Returns all divisions (Anabilim Dalı values) in the National Thesis Center
database.
Turkish: Anabilim Dalı (ABD)
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
divisions <- list_divisions()
head(divisions)
#> # A tibble: 6 x 2
#>   name                         id
#>   <chr>                        <chr>
#> 1 ACIL TIP ANABILIM DALI       1
#> 2 ADLI BILISIM ANABILIM DALI   2
\dontshow{\}) # examplesIf}
}
\seealso{
Other lookup functions: 
\code{\link{list_disciplines}()},
\code{\link{list_institutes}()},
\code{\link{list_subjects}()},
\code{\link{list_universities}()}
}
\concept{lookup functions}
