% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{list_institutes}
\alias{list_institutes}
\title{List all available institutes}
\usage{
list_institutes()
}
\value{
A tibble with two columns:
\itemize{
\item name - Character. Institute name
\item id - Character. Internal API identifier
}
}
\description{
Returns all institutes in the National Thesis Center database.
Turkish: Enstitü
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
insts <- list_institutes()
\dontshow{\}) # examplesIf}
}
\seealso{
Other lookup functions: 
\code{\link{list_disciplines}()},
\code{\link{list_divisions}()},
\code{\link{list_subjects}()},
\code{\link{list_universities}()}
}
\concept{lookup functions}
