% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{list_subjects}
\alias{list_subjects}
\title{List all available subjects}
\usage{
list_subjects()
}
\value{
A tibble with three columns:
\itemize{
\item name_tr - Character. Turkish subject name
\item name_en - Character. English subject name
\item id - Character. Internal API identifier
}
}
\description{
Returns all subjects in the National Thesis Center database.
Turkish: Konu
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
subjects <- list_subjects()
\dontshow{\}) # examplesIf}
}
\seealso{
Other lookup functions: 
\code{\link{list_disciplines}()},
\code{\link{list_divisions}()},
\code{\link{list_institutes}()},
\code{\link{list_universities}()}
}
\concept{lookup functions}
