% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{list_universities}
\alias{list_universities}
\title{List all available universities}
\usage{
list_universities()
}
\value{
A tibble with two columns:
\itemize{
\item name - Character. University name
\item id - Character. Internal API identifier
}
}
\description{
Returns all universities in the National Thesis Center database.
Turkish: Üniversite
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
unis <- list_universities()
\dontshow{\}) # examplesIf}
}
\seealso{
Other lookup functions: 
\code{\link{list_disciplines}()},
\code{\link{list_divisions}()},
\code{\link{list_institutes}()},
\code{\link{list_subjects}()}
}
\concept{lookup functions}
