% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{stats_types}
\alias{stats_types}
\title{Get thesis statistics by type}
\usage{
stats_types()
}
\value{
A tibble containing total thesis counts per type.
}
\description{
Get thesis statistics by type
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
stats <- stats_types()
stats
#> # A tibble: 1 x 4
#>   yuksek_lisans doktora tipta_uzmanlik sanatta_yeterlik
#>           <dbl>   <dbl>           <dbl>            <dbl>
#> 1        721384  178264           90030             2395
\dontshow{\}) # examplesIf}
}
\seealso{
Other statistics functions: 
\code{\link{stats_subjects}()},
\code{\link{stats_universities}()},
\code{\link{stats_years}()}
}
\concept{statistics functions}
