% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{stats_universities}
\alias{stats_universities}
\title{Get thesis statistics by university}
\usage{
stats_universities()
}
\value{
A tibble containing thesis counts per university and type.
}
\description{
Get thesis statistics by university
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
stats <- stats_universities()
\dontshow{\}) # examplesIf}
}
\seealso{
Other statistics functions: 
\code{\link{stats_subjects}()},
\code{\link{stats_types}()},
\code{\link{stats_years}()}
}
\concept{statistics functions}
