% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{stats_years}
\alias{stats_years}
\title{Get thesis statistics by year}
\usage{
stats_years()
}
\value{
A tibble containing thesis counts per year and type.
}
\description{
Get thesis statistics by year
}
\examples{
\dontrun{
stats <- stats_years()
}
}
